Remove receiptTypes: AddPage now accepts Purchase[] directly
- AddPage принимает initialPurchases: Purchase[] вместо ReceiptEdit
- Дата и магазин извлекаются из первого Purchase через $effect
- ListPage передаёт purchases напрямую в navigation.go
- Убран addPageKey и {#key} — AddPage реагирует на проп реактивно
- Удалён receiptTypes.ts
