+++ client/src/App.svelte
@@ -7,7 +7,7 @@
   import LoadingScreen from './lib/LoadingScreen.svelte';
   import ToastContainer from './lib/ToastContainer.svelte';
   import { navigation } from './lib/navigation.svelte';
-  import type { ReceiptEdit } from './lib/receiptTypes';
+  import type { Purchase } from './models/Purchase';
   import LoginPage from './pages/LoginPage.svelte';
   import AddPage from './pages/AddPage.svelte';
   import ListPage from './pages/ListPage.svelte';
@@ -17,14 +17,12 @@
   type TabId = 'add' | 'list' | 'products' | 'profile';
 
   let activeTab = $state<TabId>('add');
-  let addPageData = $state<ReceiptEdit | null>(null);
-  let addPageKey = $state(0);
+  let addPageData = $state<Purchase[]>([]);
 
   navigation.register((tab, data) => {
     activeTab = tab as TabId;
     if (tab === 'add' && data) {
-      addPageData = data as ReceiptEdit;
-      addPageKey++;
+      addPageData = data as Purchase[];
     }
   });
 
@@ -60,9 +58,7 @@
   <div class="app-shell">
     <main class="content">
       <div class="tab-panel" hidden={activeTab !== 'add'}>
-        {#key addPageKey}
-          <AddPage initialData={addPageData} />
-        {/key}
+        <AddPage initialPurchases={addPageData} />
       </div>
       <div class="tab-panel" hidden={activeTab !== 'list'}>
         <ListPage />
